requirements.txt
anthropic
python-dotenv
sentence-transformers
faiss-cpu
